Changes in v1.1:
The original select element no longer needs the 'data-styled-select-id' property.

Instead, pass the 'styled_select_id' option along when initializing the widget.

The element that displays the selected item's label is now a div that fits more fully.

The arrow's left border now will be styled the same as the replacement widget's border style.

This keeps it in line with the method that jquery.styledFileInput uses, for consistency.